SAKIII-2087
Looking at optional layouthint on tool properties to see which column the tool
should be in. There appears to be a row attribute in this as well, but Sakai2
seems to ignore it so we will do the same.

Copied the col1of2 and col2of2 styles from Sakai2 for layout when there are
2 columns, otherwise we use a regular div.

Added this layout to the trimpath template, and some logic in the javascript
to build up a different datastructure that we can pass in to the template in
order to iterate through the columns with a double loop.