SAKIII-2878
Fixing the BLTI and Sakai2Tools widgets.  One major change is that they
now literally share the same code. Previously, before the last wave of
bit rot, the BLTI widget had been copied and then modified. Since there
are so few differences I've not just put in a few switches to switch
back and forth. So even though you need to physically copy them and
change the config, there is less duped code.

Some changes to Ajax, (we make the calls ourself because the blti servlet
is a bit different than the regular data-widget ones). Added a switch
in the template to select the Sakai tool, and some i18n keys.

We need to also remember to update the hybrid nakamura config because there
have been some changes (property name changes in sak2, like imsblti ->
basiclti.sakai.property.whatever).  Also the hardcoded json config url's
in the basiclti nakamura bundle are a little problematic and need to be
search replaced for your server config currently.