Issue #2918 - Remove etherpad iframe from the DOM
This is a bugfix for IE10.

For reference, the content page could be at cam.oae.com/content/foo and 
the etherpad server is running at 0.etherpad.oae.com.

At some point during the rendering of the iframe IE mixes up the host
when loading the static assets (after exactly 10 requests actually)
The following requests are shown in the NET tab:
    (304) GET http://0.etherpad.oae.com/p/<padid>
    (304) GET http://0.etherpad.oae.com/static/css/pad.css
    (304) GET http://0.etherpad.oae.com/static/custom/pad.css
    (304) GET http://0.etherpad.oae.com/static/js/html10n.js
    ..
    (304) GET http://cam.oae.com/content/static/css/pad.css
    (304) GET http://cam.oae.com/content/static/custom/pad.css
    (304) GET http://cam.oae.com/content/static/js/html10n.js

Because the cam.oae.com/content/ URL returns the HTML for the profile page
Etherpad has no idea what to do with it and just stops processing and
ending up with the default "no access view".

I'm not entirely sure what is triggering this behaviour, but removing the
iframe from the DOM when unloading the page seems to get rid of it.
I've set min-height for the container to 500 pixels as well, so the comments
don't jump to the top of the page when removing the iframe.