Bootstrap 3 - Authentication fixes
When only one external strategy is enabled, the "Sign in" links, takes
you straight to the API endpoint.

A couple of small visual fixes are included such as:
 * aligning search icon with "Sign in"
 * proper padding in the sign in drop down
 * dotted border when in small view mode
 * new visual buttons for external strategies

As the external authentication buttons their designs have changed, the
auth-buttons.css file is no longer necessary so it has been removed.

Fixes for #3406 and 3446