Handle properly the creation and display of a loodle with the folder mechanism.
- Show the « create loodle » button in the main navigation inside a
folder.
- At the creation of a loodle, the details about the folder the loodle
is created in are send with the api call to be handle properly.
- Update the setpermissions.html file and the default.properties file
of the same module to display the correct messages when the content is
a loodle.
Update the addtofolder.html file and the default.properties file of the
same module to display the correct notification messages when the
content is a loodle.