Correction of the file default.properties shared among all the modules.
Amelioration of the creation of the sub-menu displayed in a loodle item.
Correction of the data send to the template file in the oae-esup/loodle
widget.
Update of the manifest.json file of the oae-esup/loodle widget and
amelioration of the traduction of the widget in english (by default)
and in french.